# server/openapi_server/openapi/openapi.yaml
openapi: 3.0.0
info:
  description: API definition for interacting with the back-end system.
  title: UniBar API
  version: 1.0.0
servers:
- description: Local test server.
  url: http://localhost
- description: Developement server.
  url: http://dev.unibar.xyz
- description: Production server.
  url: http://prod.unibar.xyz
paths:
  /admin/login:
    post:
      operationId: admin_login_post
      requestBody:
        $ref: '#/components/requestBodies/body_admin_login'
        content:
          application/json:
            schema:
              properties:
                admin_token:
                  format: password
                  type: string
              required:
              - admin_token
              type: object
        description: Admin login API token.
      responses:
        "200":
          description: Successfully logged in the admin.
        "400":
          description: Failed to log in the admin.
      x-openapi-router-controller: openapi_server.controllers.default_controller
    summary: Logs in the admin. Sets token in cookies.
  /admin/messages/list:
    get:
      operationId: admin_messages_list_get
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Message'
                type: array
          description: Succesfully fetched all messages.
        "400":
          description: Failed to fetch all messages.
      security:
      - AdminAuth: []
      x-openapi-router-controller: openapi_server.controllers.default_controller
    summary: Lists all messages.
  /admin/orders/list:
    get:
      operationId: admin_orders_list_get
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Order'
                type: array
          description: Succesfully fetched all orders.
        "400":
          description: Failed to fetch all orders.
      security:
      - AdminAuth: []
      x-openapi-router-controller: openapi_server.controllers.default_controller
    summary: Lists all orders.
  /admin/reports/list:
    get:
      operationId: admin_reports_list_get
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Order'
                type: array
          description: Succesfully fetched all reports.
        "400":
          description: Failed to fetch all orders.
      security:
      - AdminAuth: []
      x-openapi-router-controller: openapi_server.controllers.default_controller
    summary: Lists all user reports.
  /admin/users/list:
    get:
      operationId: admin_users_list_get
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/User'
                type: array
          description: Succesfully fetched all users.
        "400":
          description: Failed to fetch all users.
      security:
      - AdminAuth: []
      x-openapi-router-controller: openapi_server.controllers.default_controller
    summary: Lists all users.
...
  /message:
    post:
      operationId: message_post
      requestBody:
        $ref: '#/components/requestBodies/body_message'
        content:
          application/json:
            schema:
              properties:
                user_id:
                  description: Identification of a user.
                  format: uuid
                  type: string
                email:
                  description: Email format.
                  format: email
                  type: string
                name:
                  type: string
                message:
                  type: string
              required:
              - message
              - name
              type: object
        description: Send an arbitrary message to UniBar with name and message being
          optional. Email or user ID can be specified to help get back in contact
          with the message sender.
        required: true
      responses:
        "200":
          description: Successfully sent a message to UniBar.
        "400":
          description: Failed to send a message to UniBar.
      x-openapi-router-controller: openapi_server.controllers.default_controller
    summary: Send a message to UniBar.
  /order/complete:
    put:
      operationId: order_complete_put
      requestBody:
        $ref: '#/components/requestBodies/body_order_complete'
        content:
          application/json:
            schema:
              properties:
                order_id:
                  description: Identification of the order.
                  format: uuid
                  type: string
              required:
              - order_id
              type: object
        description: The ID of the order being marked as complete.
        required: true
      responses:
        "200":
          description: Successfully completed the order and transfered the delivery-token
            from food receiver to the deliverer.
        "400":
          description: Failed to complete the order and transfer the delivery-token
            from the food receiver to the deliverer.
      security:
      - UserAuth: []
      x-openapi-router-controller: openapi_server.controllers.default_controller
    summary: Completes the order and performs necessary delivery-token transfer from
      food receiver to deliverer.
  /order/claim:
    put:
      operationId: order_claim_put
      requestBody:
        $ref: '#/components/requestBodies/body_order_claim'
        content:
          application/json:
            schema:
              properties:
                order_id:
                  description: Identification of the order.
                  format: uuid
                  type: string
              required:
              - order_id
              type: object
        description: Delivery request order which the user is requesting to deliver.
        required: true
      responses:
        "200":
          description: Succesfully indicated that user intends to fulfill the order.
        "400":
          description: Could not fulfil the order.
        "403":
          description: Order is currently being fulfilled by another user.
      security:
      - UserAuth: []
      x-openapi-router-controller: openapi_server.controllers.default_controller
    summary: Indicate which order the user intends to fulfill.
  /order/report:
    post:
      operationId: order_report_post
      requestBody:
        $ref: '#/components/requestBodies/body_order_report'
        content:
          application/json:
            schema:
              properties:
                order_id:
                  description: Identification of the order.
                  format: uuid
                  type: string
                reported_id:
                  description: Identification of a user.
                  format: uuid
                  type: string
                message:
                  type: string
              required:
              - message
              - order_id
              - reported_id
              type: object
        description: Details about the order and user involved in the report. Includes
          a message.
        required: true
      responses:
        "200":
          description: Successfully reported the user.
        "400":
          description: Failed to report the user.
      security:
      - UserAuth: []
      x-openapi-router-controller: openapi_server.controllers.default_controller
    summary: Reports another user in the context of the order.
...
  /order/unclaim:
    put:
      operationId: order_unclaim_put
      requestBody:
        $ref: '#/components/requestBodies/body_order_unclaim'
        content:
          application/json:
            schema:
              properties:
                order_id:
                  description: Identification of the order.
                  format: uuid
                  type: string
              required:
              - order_id
              type: object
        description: Delivery request order which the user is no longer claiming.
        required: true
      responses:
        "200":
          description: Succesfully indicated that user no longer intends to fulfill
            the order.
        "400":
          description: Could not unclaim the order.
        "403":
          description: User is not currently claiming this order.
      security:
      - UserAuth: []
      x-openapi-router-controller: openapi_server.controllers.default_controller
    summary: Indicate which order the user no longer intends to fulfill.
  /order/{order_id}:
    get:
      operationId: order_order_id_get
      parameters:
      - description: The ID of the order which the client wants to see its info.
        explode: false
        in: path
        name: order_id
        required: false
        schema:
          $ref: '#/components/schemas/OrderID'
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
          description: Successfully fetched order.
        "400":
          description: Failed to fetch orders.
      x-openapi-router-controller: openapi_server.controllers.default_controller
    summary: Fetch a specific order.
  /orders/available:
    get:
      operationId: orders_available_get
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Order'
                type: array
          description: Successfully fetched orders available to fulfill.
        "400":
          description: Failed to fetch orders available to fulfill.
      x-openapi-router-controller: openapi_server.controllers.default_controller
    summary: Lists all orders currently available to fulfill.
...
  /user/{user_id}:
    get:
      operationId: user_user_id_get
      parameters:
      - description: User ID of the user to identify who's info to get.
        explode: false
        in: path
        name: user_id
        required: false
        schema:
          $ref: '#/components/schemas/UserID'
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: Successfully fetched user's public information.
        "400":
          description: Failed to fetch user's public information.
      security:
      - UserAuth: []
      - AdminAuth: []
      x-openapi-router-controller: openapi_server.controllers.default_controller
    summary: Get public information about the user.
  /user/{user_id}/orders/claimed:
    get:
      operationId: user_user_id_orders_claimed_get
      parameters:
      - description: User ID of which we want to see the orders they are currently claiming.
        explode: false
        in: path
        name: user_id
        required: false
        schema:
          $ref: '#/components/schemas/UserID'
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Order'
                type: array
          description: Succesfully fetched user's claimed orders.
        "400":
          description: Failed to fetch user's claimed orders.
      security:
      - UserAuth: []
      - AdminAuth: []
      x-openapi-router-controller: openapi_server.controllers.default_controller
    summary: Fetch the user's orders that they claimed.
...
  /user/{user_id}/update:
    put:
      operationId: user_user_id_update_put
      parameters:
      - description: User ID of user we are trying to update.
        explode: false
        in: path
        name: user_id
        required: false
        schema:
          $ref: '#/components/schemas/UserID'
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/body_user_update'
        content:
          application/json:
            schema:
              properties:
                name:
                  description: User's display name.
                  type: string
                password:
                  description: Password format.
                  format: password
                  type: string
                phone_number:
                  description: Phone number format.
                  pattern: ^\d{10}$
                  type: string
                etransfer_email:
                  description: Email format.
                  format: email
                  type: string
              type: object
        description: Information about the user to change. For changing the user's
          name, password, phone_number, and etransfer_email.
        required: true
      responses:
        "200":
          description: Successfully updated user information.
        "400":
          description: Failed to update user information.
        "401":
          description: Not authorized to update user's information.
      security:
      - UserAuth: []
      - AdminAuth: []
      x-openapi-router-controller: openapi_server.controllers.default_controller
    summary: Update user information.
  /users/login:
    post:
      operationId: users_login_post
      requestBody:
        $ref: '#/components/requestBodies/body_users_login'
        content:
          application/json:
            schema:
              properties:
                email:
                  description: Email format.
                  format: email
                  type: string
                password:
                  description: Password format.
                  format: password
                  type: string
              required:
              - email
              - password
              type: object
        description: Credentials of the user to log in. Email and password required
          for logging in the user.
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserID'
          description: Successfully logged in.
        "400":
          description: Failed to log in user.
      summary: Log in a user. Set cookie auth token.
      x-openapi-router-controller: openapi_server.controllers.default_controller
    summary: Logging in a user.
  /users/register:
    post:
      operationId: users_register_post
      requestBody:
        $ref: '#/components/requestBodies/body_users_register'
        content:
          application/json:
            schema:
              properties:
                email:
                  description: Email format.
                  format: email
                  type: string
                password:
                  description: Password format.
                  format: password
                  type: string
                name:
                  description: User's display name.
                  type: string
                phone_number:
                  description: Phone number format.
                  pattern: ^\d{10}$
                  type: string
              required:
              - email
              - name
              - password
              - phone_number
              type: object
        description: Specify the basic and mandatory fields to create a new user (email,
          password, name, phone_number).
        required: true
      responses:
        "200":
          description: Successfully registered a new user.
        "400":
          description: Failed to register a new user.
        "403":
          description: User already exists.
      summary: Create a new user.
      x-openapi-router-controller: openapi_server.controllers.default_controller
    summary: Registering a user.
components:
  requestBodies:
...
    body_message:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/body_message'
      required: true
    body_admin_login:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/body_admin_login'
      required: true
    body_order_complete:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/body_order_complete'
      required: true
    body_order_report:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/body_order_report'
      required: true
    body_order_claim:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/body_order_claim'
      required: true
    body_order_unclaim:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/body_order_unclaim'
...
  schemas:
...
    OrderID:
      description: Identification of the order.
      format: uuid
      type: string
...
    Message:
      description: A user message to UniBar.
      example:
        user_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        message_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        time: 2000-01-23T04:56:07.000+00:00
        message: message
        email: email
      properties:
        message_id:
          format: uuid
          type: string
        user_id:
          description: Identification of a user.
          format: uuid
          type: string
        email:
          description: Email format.
          format: email
          type: string
        message:
          type: string
        time:
          format: date-time
          type: string
      type: object
...
  securitySchemes:
    UserAuth:
      in: cookie
      name: user_token
      type: apiKey
      x-apikeyInfoFunc: openapi_server.controllers.security_controller_.info_from_UserAuth
    AdminAuth:
      in: header
      name: admin_token
      type: apiKey
      x-apikeyInfoFunc: openapi_server.controllers.security_controller_.info_from_AdminAuth